rename(waiting): move waiting to redis package #337
