+++ src/main/java/com/matdang/seatdang/waiting/repository/query/WaitingQueryRepository.java
@@ -1,6 +1,6 @@
 //package com.matdang.seatdang.waiting.repository.query;
 //
-//import com.matdang.seatdang.waiting.entity.Waiting;
+//import com.matdang.seatdang.waiting.redis.Waiting;
 //import com.matdang.seatdang.waiting.entity.WaitingStatus;
 //import com.matdang.seatdang.waiting.entity.WaitingStorage;
 //import com.matdang.seatdang.waiting.repository.query.dto.WaitingDto;
