+++ src/main/java/com/matdang/seatdang/waiting/repository/query/dto/WaitingDto.java
@@ -1,13 +1,11 @@
 package com.matdang.seatdang.waiting.repository.query.dto;
 
-import com.matdang.seatdang.waiting.entity.Waiting;
+import com.matdang.seatdang.waiting.redis.Waiting;
 import com.matdang.seatdang.waiting.entity.WaitingStatus;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
 @Data
