+++ src/test/java/com/matdang/seatdang/waiting/repository/WaitingStorageRepositoryTest.java
@@ -1,10 +1,8 @@
 package com.matdang.seatdang.waiting.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import com.matdang.seatdang.waiting.entity.CustomerInfo;
-import com.matdang.seatdang.waiting.entity.Waiting;
 import com.matdang.seatdang.waiting.entity.WaitingStatus;
 import com.matdang.seatdang.waiting.entity.WaitingStorage;
 import jakarta.persistence.EntityManager;
