+++ src/test/java/com/matdang/seatdang/waiting/service/facade/RedissonLockWaitingFacadeTest.java
@@ -6,11 +6,9 @@ import static org.mockito.Mockito.when;
 import com.matdang.seatdang.auth.service.AuthService;
 import com.matdang.seatdang.member.entity.Customer;
 import com.matdang.seatdang.waiting.dto.UpdateRequest;
-import com.matdang.seatdang.waiting.dto.WaitingId;
-import com.matdang.seatdang.waiting.entity.Waiting;
+import com.matdang.seatdang.waiting.redis.Waiting;
 import com.matdang.seatdang.waiting.entity.WaitingStatus;
 import com.matdang.seatdang.waiting.repository.WaitingRepository;
-import com.matdang.seatdang.waiting.service.WaitingCustomerService;
 import com.matdang.seatdang.waiting.service.WaitingService;
 import java.util.ArrayList;
 import java.util.List;
