+++ src/test/java/com/matdang/seatdang/waiting/service/facade/RedissonWaitingCustomerTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.when;
 
 import com.matdang.seatdang.auth.service.AuthService;
 import com.matdang.seatdang.member.entity.Customer;
-import com.matdang.seatdang.waiting.entity.Waiting;
+import com.matdang.seatdang.waiting.redis.Waiting;
 import com.matdang.seatdang.waiting.entity.WaitingStatus;
 import com.matdang.seatdang.waiting.service.WaitingService;
 import java.util.ArrayList;
